+++ scripts/babel-relay-plugin/lib/RelayQLPrinter.js
@@ -347,7 +347,7 @@ function printField(field, type, options, requisiteFields, isGenerated, parentTy
     subRequisiteFields.node = true;
   }
 
-  if (types.isAbstractType(fieldDecl.type)) {
+  if (types.isAbstractType(types.getNamedType(fieldDecl.type))) {
     metadata.dynamic = true;
     subRequisiteFields.__typename = true;
   }