+++ src/mutation/__tests__/RelayMutationQuery-test.js
@@ -191,6 +191,7 @@ describe('RelayMutationQuery', () => {
           ... on Story {
             id,
             actors {
+              __typename,
               id,
               name
             },