+++ src/traversal/__tests__/diffRelayQuery-test.js
@@ -1003,9 +1003,10 @@ describe('diffRelayQuery', () => {
     expect(diffQueries[0]).toEqualQueryRoot(getVerbatimNode(Relay.QL`
       query {
         node(id:"4808495") {
-          id,
           __typename,
+          id,
           ... on Actor {
+            __typename,
             id,
             lastName,
             name,
@@ -1017,9 +1018,10 @@ describe('diffRelayQuery', () => {
     expect(diffQueries[1]).toEqualQueryRoot(getVerbatimNode(Relay.QL`
       query {
         node(id:"1023896548") {
-          id,
           __typename,
+          id,
           ... on Actor {
+            __typename,
             firstName,
             id,
             lastName,